Move all code to chunk handler
