+++ Assets/Scripts/MineralSpawner.cs
@@ -73,7 +73,7 @@ public class MineralSpawner : MonoBehaviour
             }
         }
 
-        Debug.LogWarning("No valid terrain found for rock spawn");
+        // Debug.LogWarning("No valid terrain found for rock spawn");
         return Vector3.zero;
     }
 }
